[#1087] Split logic of Panel into a separate component.
