+++ frontend/src/commit-panel/diff/Diff.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import DiffPanel from '../../commits-table/commit-full-diff/DiffPanel.react';
+import DiffPanel from '../../commits-table/commit-full-diff/DiffPanel';
 
 interface DiffProps {
   diff: string;
